feat(home): adding pages fixing buggs
