+++ home/images/languagesandsoftware/processed_images.js
@@ -1,0 +1,25 @@
+const processed_images = [
+    '1.png',
+    '10.png',
+    '11.png',
+    '12.png',
+    '13.png',
+    '14.png',
+    '15.png',
+    '16.png',
+    '17.png',
+    '18.png',
+    '19.png',
+    '2.png',
+    '20.png',
+    '21.png',
+    '22.png',
+    '23.png',
+    '3.png',
+    '4.png',
+    '5.png',
+    '6.png',
+    '7.png',
+    '8.png',
+    '9.png',
+];
