+++ home/pages/languages-and-software.html
@@ -1,9 +1,9 @@
 <script src="../scripts/Bootstrapper.js"></script>
-<script>Bootstrapper("Luna", "luna")</script>
+<script>Bootstrapper("Languages and Software", "green")</script>
 
 <div class="content">
     <div class="catalog"></div>
 </div>
 
 <script src="../scripts/Catalog.js"></script>
-<script>LoadCatalog("../images/luna", "catalog")</script>
+<script>LoadCatalog("../images/languagesandsoftware", "catalog")</script>
