+++ home/pages/less-homework-relates-to-greater-learing.html
@@ -1,0 +1,9 @@
+<script src="../scripts/Bootstrapper.js"></script>
+<script>Bootstrapper("Less Homework Relates To Greater Learning", "white")</script>
+
+<div class="content">
+    <div class="paper"><hr>
+        <img src="../images/articles/less-homework-relates-to-greater-learing_page-1.png"><hr>
+        <img src="../images/articles/less-homework-relates-to-greater-learing_page-2.png">
+    </div>
+</div>
