+++ home/scripts/image processing/paths.json
@@ -1,0 +1,6 @@
+{
+    "paths": [
+        "../../images/luna",
+        "../../images/languagesandsoftware"
+    ]
+}
