+++ home/styles/navbar.css
@@ -14,7 +14,7 @@
     height: 75px;
 }
 
-.navbar > img {
+.navbar > div {
     background-color: var(--accent-light);
 
     margin-left: calc((75px - var(--navbar-item-height) - (2px * 2)) / 2);
@@ -25,10 +25,12 @@
     border-radius: 10px;
     border: solid var(--accent-dark) 2px;
 
+    overflow: hidden;
+
     transition-duration: 200ms;
 }
 
-.navbar > img:hover {
+.navbar > div:hover {
     background-color: var(--accent-dark);
 
     filter: drop-shadow(0 0 5px var(--accent-dark));
@@ -38,6 +40,12 @@
     width: var(--navbar-item-height-hover);
     height: var(--navbar-item-height-hover);
 }
+
+.navbar > div > img {
+    width: 100%;
+    height: 100%;
+}
+
 
 .navbar > hr {
     background-color: var(--accent-light);
