+++ home/styles/themes/white.css
@@ -1,0 +1,15 @@
+/* Pallet "https://coolors.co/palette/f8f9fa-e9ecef-dee2e6-ced4da-adb5bd-6c757d-495057-343a40-212529" */
+
+:root {
+    --accent: #CED4DA;
+    --accent-light: #DEE2E6;
+    --accent-dark: #ADB5BD;
+
+    --text: black;
+
+    --background: #E9ECEF;
+
+    font-family: "Roboto Mono", serif;
+    font-optical-sizing: auto;
+    font-style: normal;
+}
